+++ resources/views/livewire/rdelovi/komponente/rezervni-deo-histroy.blade.php
@@ -1,0 +1,5 @@
+<div>
+    <label for="lokacija">Rezervni deo:</label>
+    <livewire:rdelovi.komponente.rzervni-deo-info :partType_id="$part_id" :key="time()" />
+    
+</div>